Move unit tests to packages
